pyproject.toml
[tool.poetry]
name = "espenhk/tmnf-ai"
version = "0.1.0"
description = "Trackmania AI hobby project"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.tmnf.dependencies]
pandas = ">=2.0"
numpy = ">=1.26"
scipy = ">=1.12"
gymnasium = ">=0.29"
stable-baselines3 = ">=2.0"
pyyaml = ">=6.0"
tensorboard = ">=2.0"
matplotlib = ">=3.8"
opencv-python = ">=4.8"
mss = ">=9.0"
pywin32 = ">=306"
tminterface = ">=1.0"
pygbx = ">=0.3"

[tool.poetry.group.tmnf-test.dependencies]
pytest = ">=8.0"
numpy = ">=1.26"
scipy = ">=1.12"
pyyaml = ">=6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
